add zod for url state validation, and move boardcardelement to own file
