+++ package.json
@@ -31,6 +31,7 @@
     "microdiff": "1.5.0",
     "next": "^15.3.1",
     "next-themes": "^0.4.6",
+    "nuqs": "^2.4.3",
     "pako": "2.1.0",
     "postcss": "^8.5.3",
     "react": "^19.0.0",
@@ -38,7 +39,8 @@
     "tailwind-merge": "^3.2.0",
     "tailwindcss": "^4.1.4",
     "tw-animate-css": "^1.2.5",
-    "vite": "^6.1.1"
+    "vite": "^6.1.1",
+    "zod": "^3.24.3"
   },
   "devDependencies": {
     "@babel/plugin-syntax-import-attributes": "^7.26.0",
